DSS-3390 : fix notice numbers incorporation + add signature policy unit tests

+++ dss-cb-ades/src/main/java/eu/europa/esig/dss/cbades/signature/CBAdESLevelBaselineB.java
@@ -562,7 +562,7 @@ public class CBAdESLevelBaselineB {
     private CBORArray getNoticeNumbersArray(int[] noticeNumbers) {
         CBORArray cborArray = new CBORArray(noticeNumbers.length);
         for (int number : noticeNumbers) {
-            cborArray.add(number);
+            cborArray.add((long) number);
         }
         return cborArray;
     }
